update router using createBrowserRoute
