+++ frontend/React_Accretion/accretion/src/about-us/AboutUs.js
@@ -1,0 +1,8 @@
+
+export default function AboutUs () {
+    return (
+        <div>
+            <h1> About Us </h1>
+        </div>
+    )
+}
