+++ frontend/React_Accretion/accretion/src/layout/Header.css
@@ -1,0 +1,47 @@
+* {
+    font-family: Arial;
+    font-size: 20px;
+    color: black; 
+    text-decoration: none;
+}
+
+.header {
+    display: flex; 
+    align-items: center;
+}
+
+.header_logo {
+    width: 15vw; 
+}
+
+.menu_container {
+    padding: 3%;
+}
+
+.submenu {
+    display: flex;
+}
+
+/*remove the dropdown button arrow*/
+.dropdown-toggle::after {
+    display: none; 
+}
+
+/* modify the button color */
+#dropdown-basic {
+    background-color: none;
+    color: black; 
+    border: 0;
+    border-radius: 15px;
+}
+
+/* modify the button hover over color */
+#dropdown-basic:hover {
+    background-color: gray;
+}
+
+/* modify the Link behavior */
+.header a {
+    text-decoration: none;
+    color: black
+}
