+++ frontend/React_Accretion/accretion/src/layout/RootLayout.js
@@ -1,0 +1,14 @@
+import { Outlet } from 'react-router-dom';
+import Header from './Header';
+
+export default function RootLayout() {
+    return (
+        <div>
+            <Header /> 
+            <main>
+                < Outlet />
+            </main>
+        </div>
+    )
+
+};
